requirements.txt
numpy==1.22.4
pandas==1.4.0
scikit-learn
torch==2.0.0
joblib
rdkit
pyyaml
addict
tqdm
transformers
pymatgen
huggingface_hub
# PyTorch Geometric
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
